Fix de seleccion de hora reestablese a hoy la fecha seleccionada.
